git-svn-id: https://svn.jboss.org/repos/labs/labs/jbossrules/branches/DroolsChance@26839 c60d74c8-e8f6-0310-9e8f-d4a2fc68ab70

+++ drools-core/src/main/java/org/drools/degrees/operators/IDiscountOperator.java
@@ -4,6 +4,6 @@ import org.drools.degrees.IDegree;
 
 public interface IDiscountOperator extends IDegreeCombiner {
 	
-	public IDegree discount(IDegree arg, float strength);
+	public IDegree discount(IDegree arg, double strength);
 
 }
