+++ drools-core/src/main/java/org/drools/degrees/operators/intervals/IntervalForallOperator.java
@@ -13,8 +13,8 @@ public class IntervalForallOperator implements IDegreeCombiner {
 	public IDegree eval(IDegree[] args, IDegreeFactory factory) {
 		if (args == null || args.length == 0)
 			return factory == null ? IntervalDegreeFactory.UNKNOWN : factory.Unknown();		
-		float tau = 1.0f;
-		float phi = 0.0f;
+		double tau = 1.0f;
+		double phi = 0.0f;
 		for (IDegree deg : args) {
 			IntervalDegree ival = (IntervalDegree) deg;
 			if (ival == null) 
@@ -28,8 +28,8 @@ public class IntervalForallOperator implements IDegreeCombiner {
 	public IDegree eval(Collection<? extends IDegree> args, IDegreeFactory factory) {
 		if (args == null || args.size() == 0)
 			return factory == null ? IntervalDegreeFactory.UNKNOWN : factory.Unknown();		
-		float tau = 1.0f;
-		float phi = 0.0f;
+		double tau = 1.0f;
+		double phi = 0.0f;
 		for (IDegree deg : args) {
 			IntervalDegree ival = (IntervalDegree) deg;
 			if (ival == null)
