+++ drools-core/src/main/java/org/drools/degrees/operators/intervals/IntervalMaxOrOperator.java
@@ -13,8 +13,8 @@ public class IntervalMaxOrOperator implements IDegreeCombiner {
 	public IDegree eval(IDegree[] args, IDegreeFactory factory) {
 		if (args == null || args.length == 0)
 			return factory == null ? IntervalDegreeFactory.UNKNOWN : factory.Unknown();		
-		float tau = 0;
-		float phi = 1.0f;
+		double tau = 0;
+		double phi = 1.0f;
 		for (IDegree deg : args) {
 			IntervalDegree ival = (IntervalDegree) deg;
 			if (ival == null) 
@@ -28,8 +28,8 @@ public class IntervalMaxOrOperator implements IDegreeCombiner {
 	public IDegree eval(Collection<? extends IDegree> args, IDegreeFactory factory) {
 		if (args == null || args.size() == 0)
 			return factory == null ? IntervalDegreeFactory.UNKNOWN : factory.Unknown();		
-		float tau = 0;
-		float phi = 1.0f;
+		double tau = 0;
+		double phi = 1.0f;
 		for (IDegree deg : args) {
 			IntervalDegree ival = (IntervalDegree) deg;
 			if (ival == null)
